苏东方

+++ app/src/main/java/com/yangbo/maoyan1/bean/DaiYIngRcViewBean.java
@@ -115,7 +115,7 @@ public class DaiYIngRcViewBean {
             private String img;
             private boolean late;
             private int localPubSt;
-            private int mk;
+            private double mk;
             private String nm;
             private int pn;
             private boolean preShow;
@@ -124,7 +124,7 @@ public class DaiYIngRcViewBean {
             private int recentShowDate;
             private int recentShowNum;
             private String rt;
-            private int sc;
+            private double sc;
             private String scm;
             private String showInfo;
             private int showNum;
@@ -244,11 +244,11 @@ public class DaiYIngRcViewBean {
                 this.localPubSt = localPubSt;
             }
 
-            public int getMk() {
+            public double getMk() {
                 return mk;
             }
 
-            public void setMk(int mk) {
+            public void setMk(double mk) {
                 this.mk = mk;
             }
 
@@ -316,11 +316,11 @@ public class DaiYIngRcViewBean {
                 this.rt = rt;
             }
 
-            public int getSc() {
+            public double getSc() {
                 return sc;
             }
 
-            public void setSc(int sc) {
+            public void setSc(double sc) {
                 this.sc = sc;
             }
